+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 - Parsley now relies on `input` events instead of `change` and `keyup` events.
+  Custom triggers are unbound once a field has failed.
 - Parsley updates the UI before firing success/error/validated events
 - Deprecated uses of ParsleyUI. Methods are now instance methods of ParsleyField,
   with modernized interfaces.